+++ libs/web/src/controls/select-input/select-input.component.ts
@@ -24,17 +24,17 @@ export class SelectInputComponent extends BaseComponent {
   inputElement: ElementRef;
 
   @Input()
-  debounceTime?: number;
+  debounceTime: number;
   @Output()
   onChangeInputValue: EventEmitter<string> = new EventEmitter<string>();
   @Output()
   onInputFocus: EventEmitter<string> = new EventEmitter<string>();
   @Input()
-  labelClass?= 'control-label';
+  labelClass = 'control-label';
   @Input()
-  inputClass?= 'form-control';
+  inputClass = 'form-control';
   @Input()
-  inputFrameClass?= '';
+  inputFrameClass = '';
   @Input()
   inFormGroup = true;
   @Input()
@@ -73,7 +73,12 @@ export class SelectInputComponent extends BaseComponent {
       return this.dataSource;
     }
     return this._items.map(item => {
-      item[this.inputTitleField] = this.getInputTitle(item);
+      // todo: refactor
+      try {
+        item[this.inputTitleField] = this.getInputTitle(item);
+      } catch (error) {
+
+      }
       return item;
     });
   }